DBZ-1491 Refactor after review
